RedirectAllTrafficTo and StatusOK helpers
* RedirectAllTrafficTo is a helper that creates a handler to
perform 301 Permanent Redirects to the specified host for any
path(route) for any traffic that is encountered.
Sample usage in production at orijtech.com is:
```go
httpsRedirectHandler := otils.RedirectAllTrafficTo("https://orijtech.com")
if err := http.ListenAndServe(":80", httpsRedirectHandler); err != nil {
    log.Fatal(err)
}
```
to redirect any non-https traffic from:
+ http://orijtech.com/*
+ http://www.orijtech.com/*

to
+ https://orijtech.com/*

* StatusOK returns true if a status code is in the 200s range ie 2XX
from 200 to 299. It is a common pattern when checking the status of
an HTTP request for success.