ToURLValues: recognize false as a blank value
Fixes #8.

As mentioned in the issue, I encountered an API that requires
boolean values for some keys in the query string.
However, merely including `false` makes some APIs treat the
value as included hence `true`.

This change makes the code recognize `false` as a blank value
hence excluding it from the values map.